refactor(components): add services
переписал запрос подсказок - services/suggestionsApi

МЕТА ДАННЫЕ: refactor 05
